+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # 0.1.0 (Unreleased)
+
+- Define the available types of packet in the `PacketType` enum.